% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_h5_data}
\alias{create_h5_data}
\title{Function to create an HDF5 dataset}
\usage{
create_h5_data(h5filename, R, betahat, se, chr = NULL)
}
\arguments{
\item{h5filename}{output file name}

\item{betahat}{a length p vector specifying the univariate regression coefficients from the association study}

\item{se}{a length p vector specifying the univariate standard errors from the association study}

\item{SiRiS}{a sparse matrix of class `dgCMatrix`, representing the LD matrix}

\item{sigma_beta}{a length p vector specifiying the current value(s) of the hyper-parameter sigma_beta}

\item{logodds}{a length p vector specifiying the current value(s) of sigma_beta}
}
\description{
Function to create an HDF5 dataset
}
