% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{normalizeLogWeights}
\alias{normalizeLogWeights}
\title{Normalize log weights}
\usage{
normalizeLogWeights(lnZ, na.rm = T)
}
\arguments{
\item{lnZ}{unnormalized log probability
normalizelogweights takes as input an array of unnormalized}
}
\description{
Normalize log weights
}
