% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prep_list}
\alias{prep_list}
\title{Helper function  to specify data and hyperparameters for the scenario}
\usage{
prep_list(x, default, chunk = NULL, chunk_max = NULL)
}
\description{
Helper function  to specify data and hyperparameters for the scenario
}
