% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5fun.R
\name{prep_rss}
\alias{prep_rss}
\title{Flexible function for specifying data and hyperparameters for RSS}
\usage{
prep_rss(datafile = NULL, options = list(), chunk = NULL,
  tot_chunks = NULL)
}
\arguments{
\item{datafile}{path to HDF5 formatted file with betahat, se, and R}

\item{options}{list containing the remaining parameters that have already been specified.}
}
\description{
Flexible function for specifying data and hyperparameters for RSS
}
\details{
The elements of options are :
betahat: vector of length p pre-specifying betahat (in this case, don't read betahat from datafile)
se: vector of length p pre-specifying se (in this case, don't read se from datafile)
R: A sparse matrix of class dgCMatrix, specifying LD (note, this is distinct from the SiRiS matrix)
SiRiS: A sparse matrix of class dgCMatrix, specifying the SiRiS matrix
tolerance: the tolerance for convergence of the VB algorithm
maxiter: the maximum number of iterations to perform
alphafile file containing initial values of alpha (txt file)
mu file containing initial values of alpha (txt file)
alpha: a length p vector specifying initial values of alpha
mu: a length p vector specifying initial values of mu
Rfile: HDF5 file to use instead of datafile for LD matrix
betahatfile: HDF5 file to use instead of datafile for betahat vector
sefile: HDF5 file to use instead of datafile for se vector
Rpath: path in HDF5 file to look for R matrix (may be name of group in case of sparse matrix or name of data in case of dense matrix)
betahatpath: path in HDF5 file to look for betahat vector 
sepath: path in HDF5 file to look for se vector
}
