% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5fun.R
\name{read_ccs_h5}
\alias{read_ccs_h5}
\alias{read_vec}
\alias{write_vec}
\alias{gen_SiRSi}
\title{read a sparse matrix from HDF5
read a sparse matrix from HDF5 stored in Compressed Column Storage (CCS) format}
\usage{
read_ccs_h5(h5filename, groupname, dataname = "data", iname = "ir",
  pname = "jc", isSymmetric = NULL)

read_vec(h5filename, datapath)

write_vec(h5filename, vec, datapath)

gen_SiRSi(h5filename, Rpath = NULL, se = NULL, check = T)
}
\arguments{
\item{h5filename}{path to HDF5 file}

\item{groupname}{name of group in the HDF5 file where the data is stored}

\item{dataname}{name of the dataset in the HDF5 file}

\item{datapath}{concatenation of groupname and data name, separated by '/'}
}
\description{
read a sparse matrix from HDF5
read a sparse matrix from HDF5 stored in Compressed Column Storage (CCS) format
}
\section{Functions}{
\itemize{
\item \code{read_vec}: Read vector from an HDF5 file

\item \code{write_vec}: Read vector from an HDF5 file

\item \code{gen_SiRSi}: Read in R and standard error, and generate the SiRiS matrix
}}

