% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_varbvsr_bigmem_squarem.R
\name{rss_varbvsr_bigmem_squarem}
\alias{rss_varbvsr_bigmem_squarem}
\title{RSS
Main function for running RSS in serial (with SQUAREM updates)}
\usage{
rss_varbvsr_bigmem_squarem(datafiles, sigb = 0.058, logodds = -2.9/log(10),
  options = list())
}
\arguments{
\item{datafiles}{a vector of HDF5 files with one file per chromosome.(See details for constraints on the layout of these files)}

\item{sigb}{The prior SD of the regression coefficients (if included), scalar}

\item{logodds}{logodds: the prior log-odds (i.e. log(prior PIP/(1-prior PIP))) of inclusion for each SNP, p by 1}

\item{options}{a list containing optional arguments (including initial values for mu and alpha)}
}
\value{
A list with the following elements:
lnZ: scalar, the variational lower bound of the marginal log likelihood (up to some constant)
      alpha: p by 1, variational estimates of the posterior inclusion probabilities 
     mu: p by 1, posterior means of the additive effects (given snp included)
     s: p by 1, posterior variances of the additive effects (given snp included)
      info: list with following fields 
              - iter: integer, number of iterations
              - maxerr: the maximum relative difference between the parameters at the last two iterations
              - sigb: scalar, the maximum likelihood estimate of sigma_beta
              - loglik: iter by 1, the variational lower bound at each iteration
}
\description{
RSS
Main function for running RSS in serial (with SQUAREM updates)
}
