% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rss_varbvsr_iter_naive_reference}
\alias{rss_varbvsr_iter_naive_reference}
\title{Single update of RSS with variational method
This function is a very close translation of the original implementation of RSS. It is kept here for testing purposes
It performs a single update}
\usage{
rss_varbvsr_iter_naive_reference(SiRiS, sigma_beta, logodds, betahat, se,
  alpha0, mu0, SiRiSr0, reverse)
}
\description{
Single update of RSS with variational method
This function is a very close translation of the original implementation of RSS. It is kept here for testing purposes
It performs a single update
}
