% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rss_varbvsr_squarem}
\alias{rss_varbvsr_squarem}
\title{Run RSS with the variational bayes algorithm accelerated with SQUAREM}
\usage{
rss_varbvsr_squarem(SiRiS, sigma_beta, logodds, betahat, se, talpha0, tmu0,
  tSiRiSr0, tolerance, itermax, verbose, lnz_tol)
}
\arguments{
\item{SiRiS}{a sparse matrix of class `dgCMatrix`, representing the LD matrix}

\item{sigma_beta}{a length p vector specifiying the current value(s) of the hyper-parameter sigma_beta}

\item{logodds}{a length p vector specifiying the current value(s) of sigma_beta}

\item{betahat}{a length p vector specifying the univariate regression coefficients from the association study}

\item{se}{a length p vector specifying the univariate standard errors from the association study}

\item{talpha0}{a length p vector specifying the initial value of alpha}

\item{tmu0}{a length p vector specifying the initial value of mu}

\item{SiRiSr0}{a length p vector specifying the initial value of SiRiSr}
}
\description{
Run RSS with the variational bayes algorithm accelerated with SQUAREM
}
