% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rss_varbvsr_squarem_fit_logodds}
\alias{rss_varbvsr_squarem_fit_logodds}
\title{Estimate hyperparameters of RSS using the EM algorithm}
\usage{
rss_varbvsr_squarem_fit_logodds(SiRiS, sigma_beta, logodds0, betahat, se,
  talpha0, tmu0, tSiRiSr0, tolerance, itermax, verbose, lnz_tol)
}
\arguments{
\item{SiRiS}{a sparse matrix of class `dgCMatrix`, representing the LD matrix}

\item{sigma_beta}{a single value (numeric) specifying the value of the hyper-parameter sigma_beta}

\item{betahat}{a length p vector specifying the univariate regression coefficients from the association study}

\item{se}{a length p vector specifying the univariate standard errors from the association study}

\item{talpha0}{a length p vector specifying the initial value of alpha}

\item{tmu0}{a length p vector specifying the initial value of mu}

\item{logodds}{a single value (numeric) specifying the value of hyper-parameter logodds}

\item{SiRiSr0}{a length p vector specifying the initial value of SiRiSr}
}
\description{
Estimate hyperparameters of RSS using the EM algorithm
}
