% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rss_varbvsr_squarem_iter}
\alias{rss_varbvsr_squarem_iter}
\title{Run RSS with the variational bayes algorithm accelerated with SQUAREM, only returning the lower bound}
\usage{
rss_varbvsr_squarem_iter(SiRiS, sigma_beta, logodds, betahat, se, talpha0, tmu0,
  tSiRiSr0, tolerance, itermax, lnz_tol)
}
\arguments{
\item{SiRiS}{a sparse matrix of class `dgCMatrix`, representing the LD matrix}

\item{sigma_beta}{a single value (numeric) specifying the value of the hyper-parameter sigma_beta}

\item{logodds}{a single value (numeric) specifying the value of hyper-parameter logodds}

\item{betahat}{a length p vector specifying the univariate regression coefficients from the association study}

\item{se}{a length p vector specifying the univariate standard errors from the association study}

\item{talpha0}{a length p vector specifying the initial value of alpha}

\item{tmu0}{a length p vector specifying the initial value of mu}

\item{SiRiSr0}{a length p vector specifying the initial value of SiRiSr}
}
\description{
Run RSS with the variational bayes algorithm accelerated with SQUAREM, only returning the lower bound
}
