\name{RSSR-package}
\alias{RSSR-package}
\alias{RSSR}
\docType{package}
\title{
\packageTitle{RSSR}
}
\description{
\packageDescription{RSSR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RSSR}
\packageIndices{RSSR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{RSSR}

Maintainer: \packageMaintainer{RSSR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
