% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rssr.R
\name{rssr_grid}
\alias{rssr_grid}
\title{Run Regression with Summary Statistics over a grid of hyperparamter values}
\usage{
rssr_grid(R, betahat, se, sigma_beta = NULL, logodds = NULL,
  options = list(use_squarem = T, use_lnztol = T, tolerance = 0.001, itermax =
  200, parallel = F, nthreads = 1))
}
\arguments{
\item{betahat}{a length p vector specifying the univariate regression coefficients from the association study.}

\item{se}{a length p vector specifying the univariate standard errors from the association study}

\item{sigma_beta}{Hyperparameter sa is the prior variance of regression 
coefficients for variables that are included in the model. If it is not provided as input, then it is set
to an evenly spaced grid (log-scale) of values from 10^0.}

\item{logodds}{Hyperparameter logodds is the prior log-odds that a variable is included in the regression model;
it is defined as logodds = log(q/(1-q)), where q is the prior probability that a variable is included in the
regression model. Note that we use the natural logarithm instead of the base-10 logarithm.
If logodds is not provided as input, then it is set to an evenly spaced grid of values from `-14` to `-4.6`}
}
\description{
This is an optimized implementation of RSS. Given a correlation matrix
 and univariate summary statistics, it will return a `data.frame` with the 
 variational lower bound evaluated over a grid of hyperparameter values.
 
 @param R Correlation matrix (`p` by `p`)
   This can be a standard dense R matrix, or a sparse matrix from the package `Matrix`
}
