% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_varbvsr_bigmem_squarem.R
\name{squarem_update}
\alias{squarem_update}
\title{Convenience function to call two RSS updates}
\usage{
squarem_update(SiRiS, sigma_beta, logodds, betahat, se, o_mat, reverse)
}
\arguments{
\item{SiRiS}{see documentation for `rss_varbvsr_bigmem_squarem`}

\item{sigma_beta}{see documentation for `rss_varbvsr_bigmem_squarem`}

\item{logodds}{see documentation for `rss_varbvsr_bigmem_squarem`}

\item{betahat}{univariate regression coefficients}

\item{se}{univariate standard errors}

\item{o_mat}{a p by 3 matrix consisting of the concatenation of the three vectors: alpha0,mu0 and SiRiSr0, the initial values of alpha,mu and SiRiSr.}
}
\description{
Convenience function to call two RSS updates
}
