% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5fun.R
\name{write_ccs_h5}
\alias{write_ccs_h5}
\title{write a sparse matrix to HDF5
write a sparse matrix to HDF5 stored in the Compressed Column Storage (CCS) format)}
\usage{
write_ccs_h5(h5filename, spmat, groupname, dataname = "data", iname = "ir",
  pname = "jc", compression_level = 8)
}
\arguments{
\item{h5filename}{path to HDF5 file}

\item{groupname}{name of group in the HDF5 file where the data is stored}

\item{dataname}{name of the dataset in the HDF5 file}

\item{datapath}{concatenation of groupname and data name, separated by '/'}
}
\description{
write a sparse matrix to HDF5
write a sparse matrix to HDF5 stored in the Compressed Column Storage (CCS) format)
}
