% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_reference_imp.R
\name{squarem_update}
\alias{squarem_update}
\title{Convenience function to call two RSS updates}
\usage{
squarem_update(SiRiS, sigma_beta, logodds, betahat, se, o_mat, reverse)
}
\arguments{
\item{SiRiS}{a sparse matrix of class `dgCMatrix`, representing the LD matrix}

\item{sigma_beta}{a single value (numeric) specifying the value of the hyper-parameter sigma_beta}

\item{logodds}{a single value (numeric) specifying the value of hyper-parameter logodds}

\item{betahat}{a length p vector specifying the univariate regression coefficients from the association study}

\item{se}{a length p vector specifying the univariate standard errors from the association study}

\item{o_mat}{a p by 3 matrix consisting of the concatenation of the three vectors: alpha0,mu0 and SiRiSr0, the initial values of alpha,mu and SiRiSr.}
}
\description{
Convenience function to call two RSS updates
}
